Working on parse functions
The main task of this week is to implement the key algorithms in
`ParallelParseFunctions` and `SerialParseFunctions`.
`ParallelParseFunctions` should adopt an aggressive strategy to provide
the longest matching path for token.
`SerialParseFunctions`, on the other hand, will behave more
conservatively: it adopts a stack-based fallback strategy and will
choose to fallback to the previous "step" when encountering a failed
match.

I implemented the `Bitset` to record which items have been executed,
and it will be skipped during fallback.

At the same time, I made some modifications to the definitions in
`config.py` to prevent infinite recursive execution.

History:

CI: Remove `Debug` build in CI.

[WIP] Generate the parse functions using
`config.py`.

CI: Using clang++ instead of g++.

[WIP]: Fix the issue where non-flat data
structures are not correctly calling the constructor function.

[WIP] Remove useless lex actions.

[WIP] Fix recursive gram definitions.

[WIP] Make `SerialParseFunctions` try again until
it's `ParseFunctionOutputs` elements run out of repeat times.

[WIP] Move `ParallelParseFunctions` def etc. to
class members.

[WIP] Split the large `function_impl.h` into small
`cc` files.

[WIP] Working on parse functions.